[REF] CF: split side panel into multiple components
The CF side panel was huge and complex, it was a single component that
handled absolutely everything. This commit split the panel into a component
to display the CFs preview, and one to edit a CF.

For future work, it is most likely possible to split the CF editor into
multiple components, one for each CF type. The HTML/CSS of the preview
component could also be improved to use flexes rather than absolute
positioning everywhere.

Task: 3504521
