+++ src/components/side_panel/conditional_formatting/cf_editor/cell_is_rule_editor.xml
