+++ src/components/side_panel/conditional_formatting/cf_editor/color_scale_rule_editor.xml
