+++ src/components/side_panel/conditional_formatting/cf_editor/icon_set_rule_editor.xml
