+++ src/components/side_panel/conditional_formatting/cf_preview_list/cf_preview_list.xml
@@ -1,0 +1,85 @@
+<templates>
+  <t t-name="o-spreadsheet-ConditionalFormatPreviewList">
+    <div class="o-cf-preview-list">
+      <div
+        t-on-click="(ev) => props.onPreviewClick(cf)"
+        t-foreach="props.conditionalFormats"
+        t-as="cf"
+        t-key="cf.id">
+        <t t-call="o-spreadsheet-ConditionalFormattingPanelPreview"/>
+      </div>
+      <t t-if="state.mode === 'list'">
+        <div
+          class="btn btn-link o-sidePanel-btn-link o-cf-add"
+          t-on-click.prevent.stop="props.onAddConditionalFormat">
+          + Add another rule
+        </div>
+        <div
+          class="btn btn-link o-sidePanel-btn-link o-cf-reorder"
+          t-on-click="reorderConditionalFormats">
+          Reorder rules
+        </div>
+      </t>
+      <t t-else="">
+        <div class="btn btn-link o-sidePanel-btn-link o-cf-exit-reorder" t-on-click="stopReorder">
+          Stop reordering rules
+        </div>
+      </t>
+    </div>
+  </t>
+
+  <t t-name="o-spreadsheet-ConditionalFormattingPanelPreview">
+    <div class="o-cf-preview d-flex" t-att-class="{ 'o-cf-cursor-ptr': state.mode !== 'reorder' }">
+      <t t-if="cf.rule.type==='IconSetRule'">
+        <div class="o-cf-preview-icon d-flex justify-content-around align-items-center me-2">
+          <t t-call="o-spreadsheet-Icon.{{icons[cf.rule.icons.upper].template}}"/>
+          <t t-call="o-spreadsheet-Icon.{{icons[cf.rule.icons.middle].template}}"/>
+          <t t-call="o-spreadsheet-Icon.{{icons[cf.rule.icons.lower].template}}"/>
+        </div>
+      </t>
+      <t t-else="">
+        <div
+          t-att-style="getPreviewImageStyle(cf.rule)"
+          class="o-cf-preview-icon d-flex justify-content-around align-items-center me-2">
+          123
+        </div>
+      </t>
+      <div class="o-cf-preview-description">
+        <div class="o-cf-preview-ruletype">
+          <div class="o-cf-preview-description-rule text-truncate">
+            <t t-esc="getDescription(cf)"/>
+          </div>
+        </div>
+        <div class="o-cf-preview-range text-truncate" t-esc="cf.ranges"/>
+      </div>
+      <t t-if="state.mode === 'reorder'">
+        <div class="o-cf-reorder">
+          <t t-if="!cf_first">
+            <div
+              class="o-cf-reorder-button-up o-cf-reorder-button"
+              t-on-click.stop="(ev) => this.reorderRule(cf, 'up', ev)">
+              <t t-call="o-spreadsheet-Icon.CARET_UP"/>
+            </div>
+          </t>
+          <t t-if="!cf_last">
+            <div
+              class="o-cf-reorder-button-down o-cf-reorder-button"
+              t-on-click.stop="(ev) => this.reorderRule(cf, 'down', ev)">
+              <t t-call="o-spreadsheet-Icon.CARET_DOWN"/>
+            </div>
+          </t>
+        </div>
+      </t>
+      <t t-else="">
+        <div class="o-cf-delete">
+          <div
+            class="o-cf-delete-button text-muted"
+            t-on-click.stop="(ev) => this.deleteConditionalFormat(cf, ev)"
+            title="Remove rule">
+            <t t-call="o-spreadsheet-Icon.TRASH"/>
+          </div>
+        </div>
+      </t>
+    </div>
+  </t>
+</templates>
